+++ src/components/molecules/ReactListV1/component/reactListView.tsx
@@ -70,6 +70,10 @@ import { autoDetailsList } from '../atoms/detailsList';
 import { findParentElementPropLikeThis } from '../../../../logic/DOM/Search/domSearch';
 import { updateReactListItem } from '../functions/updateReactListItem';
 import { doesObjectExistInArray } from '../../../../logic/Arrays/searching/objectfind';
+
+// import { createPanelButtonsV1 } from '../atoms/createPanelButtonsV1';
+import { createPanelButtonsV2 } from '../atoms/createPanelButtonsV2';
+
 require('@mikezimm/fps-styles/dist/reactListView.css');
 
 
@@ -124,135 +128,135 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
 
     private delim = '|||';
 
-    /**
-     * 
-     * @param quickCommands 
-     * @param item 
-     * @param sourceUserInfo  //This is just passed in in order to allow for user targeted b.showWhenEvalTrue checks
-     */
-    private createPanelButtons ( quickCommands: IQuickCommandsDesign, item: IDrillItemInfo , sourceUserInfo: IUser ) {
+    // /**
+    //  * 
+    //  * @param quickCommands 
+    //  * @param item 
+    //  * @param sourceUserInfo  //This is just passed in in order to allow for user targeted b.showWhenEvalTrue checks
+    //  */
+    // private createPanelButtons ( quickCommands: IQuickCommandsDesign, item: IDrillItemInfo , sourceUserInfo: IUser ) {
 
-        let allButtonRows : any[] = [];
+    //     let allButtonRows : any[] = [];
 
-        //Adjusted per:  https://github.com/mikezimm/drilldown7/issues/211
-        if ( !quickCommands || !quickCommands.buttons || quickCommands.buttons.length === 0 ) { 
-            return NoCommandsInfo;
+    //     //Adjusted per:  https://github.com/mikezimm/drilldown7/issues/211
+    //     if ( !quickCommands || !quickCommands.buttons || quickCommands.buttons.length === 0 ) { 
+    //         return NoCommandsInfo;
 
-        } else {
+    //     } else {
 
-            let buildAllButtonsTest = true;
-            if ( quickCommands.showWhenEvalTrue && quickCommands.showWhenEvalTrue.length > 0 ) {
+    //         let buildAllButtonsTest = true;
+    //         if ( quickCommands.showWhenEvalTrue && quickCommands.showWhenEvalTrue.length > 0 ) {
 
-                //2022-01-18:  Added Try catch when testing and found my typed in quick command had error.
-                try {
-                    buildAllButtonsTest = eval( quickCommands.showWhenEvalTrue );
+    //             //2022-01-18:  Added Try catch when testing and found my typed in quick command had error.
+    //             try {
+    //                 buildAllButtonsTest = eval( quickCommands.showWhenEvalTrue );
                 
-                    if ( buildAllButtonsTest === true ) {
-                        //build all the buttons ( subject to individual button checks )
-                    } else { buildAllButtonsTest = false; }
-                } catch (e) {
-                    buildAllButtonsTest = false;
-                    let errMessage = getHelpfullError(e, false, false);
-                    console.log(`ERROR:  createPanelButtons: quickCommands.showWhenEvalTrue !!!`, quickCommands.showWhenEvalTrue);
-                    console.log(`ERROR:  createPanelButtons: quickCommands.showWhenEvalTrue Error Details`, errMessage);
+    //                 if ( buildAllButtonsTest === true ) {
+    //                     //build all the buttons ( subject to individual button checks )
+    //                 } else { buildAllButtonsTest = false; }
+    //             } catch (e) {
+    //                 buildAllButtonsTest = false;
+    //                 let errMessage = getHelpfullError(e, false, false);
+    //                 console.log(`ERROR:  createPanelButtons: quickCommands.showWhenEvalTrue !!!`, quickCommands.showWhenEvalTrue);
+    //                 console.log(`ERROR:  createPanelButtons: quickCommands.showWhenEvalTrue Error Details`, errMessage);
 
-                    alert(`createPanelButtons: quickCommands.showWhenEvalTrue error !!! Check the console for details:   ${quickCommands.showWhenEvalTrue}`);
-                }
+    //                 alert(`createPanelButtons: quickCommands.showWhenEvalTrue error !!! Check the console for details:   ${quickCommands.showWhenEvalTrue}`);
+    //             }
 
-            }
+    //         }
 
-            if ( buildAllButtonsTest === true ) {
-                quickCommands.buttons.map( (buttonRow, r) => {
+    //         if ( buildAllButtonsTest === true ) {
+    //             quickCommands.buttons.map( (buttonRow, r) => {
 
-                    if ( buttonRow && buttonRow.length > 0 ) {
-                      let rowResult : any = null;
-                      const buttons : any[] = [];
+    //                 if ( buttonRow && buttonRow.length > 0 ) {
+    //                   let rowResult : any = null;
+    //                   const buttons : any[] = [];
     
-                        buttonRow.map( (b: IQuickButton,i: number) => {
+    //                     buttonRow.map( (b: IQuickButton,i: number) => {
     
-                            let buildThisButton = true;
+    //                         let buildThisButton = true;
     
-                            /**
-                             * showWhenEvalTrue must be run in the context of this section of code to be valid.
-                             */
+    //                         /**
+    //                          * showWhenEvalTrue must be run in the context of this section of code to be valid.
+    //                          */
     
-                            if ( b.showWhenEvalTrue && b.showWhenEvalTrue.length > 0 ) {
+    //                         if ( b.showWhenEvalTrue && b.showWhenEvalTrue.length > 0 ) {
 
-                                //2022-01-18:  Added Try catch when testing and found my typed in quick command had error.
-                                try {
-                                  const buildButtonTest = eval( b.showWhenEvalTrue );
-                                    if ( buildButtonTest === true ) {
-                                        //build all the buttons
-                                    } else { buildThisButton = false; }
-                                } catch (e) {
-                                  const errMessage = getHelpfullError(e, false, false);
-                                    console.log(`createPanelButtons: b[${i}].showWhenEvalTrue error !!!`, b.showWhenEvalTrue);
-                                    console.log(`createPanelButtons: b[${i}].showWhenEvalTrue Error Details`, errMessage);
+    //                             //2022-01-18:  Added Try catch when testing and found my typed in quick command had error.
+    //                             try {
+    //                               const buildButtonTest = eval( b.showWhenEvalTrue );
+    //                                 if ( buildButtonTest === true ) {
+    //                                     //build all the buttons
+    //                                 } else { buildThisButton = false; }
+    //                             } catch (e) {
+    //                               const errMessage = getHelpfullError(e, false, false);
+    //                                 console.log(`createPanelButtons: b[${i}].showWhenEvalTrue error !!!`, b.showWhenEvalTrue);
+    //                                 console.log(`createPanelButtons: b[${i}].showWhenEvalTrue Error Details`, errMessage);
 
-                                    alert(`createPanelButtons: quickCommands.showWhenEvalTrue error !!! Check the console for details:   ${quickCommands.showWhenEvalTrue}`);
-                                }
+    //                                 alert(`createPanelButtons: quickCommands.showWhenEvalTrue error !!! Check the console for details:   ${quickCommands.showWhenEvalTrue}`);
+    //                             }
 
-                            }
+    //                         }
 
-                            if ( buildThisButton === true ) {
+    //                         if ( buildThisButton === true ) {
 
-                              const buttonStyles: React.CSSProperties = b.styleButton ? b.styleButton as React.CSSProperties :  { minWidth: buttonRow.length === 1 ? '350px' : '', padding: '25px', marginBottom: '10px', fontSize: 'larger' };
+    //                           const buttonStyles: React.CSSProperties = b.styleButton ? b.styleButton as React.CSSProperties :  { minWidth: buttonRow.length === 1 ? '350px' : '', padding: '25px', marginBottom: '10px', fontSize: 'larger' };
 
-                              //Tried adding 
-                              // const IconElement = b.icon ? <Icon iconName= { 'Emoji2' } style={ defaultBannerCommandStyles }/> : undefined;
-                              const buttonID = ['ButtonID', r, i , item.Id].join(this.delim);
-                              const buttonTitle = b.label;
-                              // const iconName: string = b.icon;
-                              const buttonIcon : IIconProps = { iconName: b.icon, style: defaultBannerCommandStyles }
-                              const thisButton = b.primary === true ?
+    //                           //Tried adding 
+    //                           // const IconElement = b.icon ? <Icon iconName= { 'Emoji2' } style={ defaultBannerCommandStyles }/> : undefined;
+    //                           const buttonID = ['ButtonID', r, i , item.Id].join(this.delim);
+    //                           const buttonTitle = b.label;
+    //                           // const iconName: string = b.icon;
+    //                           const buttonIcon : IIconProps = { iconName: b.icon, style: defaultBannerCommandStyles }
+    //                           const thisButton = b.primary === true ?
 
-                               //Tried adding  iconName into Primary Button, does not work,
-                               //Tried adding IconElement into icon Props, can't see it.
-                               //Tried adding Icon Element into the div next to Primary button and could see it.
-                               //
-                                    <div id={ buttonID } title={ buttonTitle } >
-                                      <PrimaryButton style= { buttonStyles } iconProps= { buttonIcon } text={b.label} onClick={this._panelButtonClicked.bind(this)} disabled={b.disabled} checked={b.checked} /></div>:
+    //                            //Tried adding  iconName into Primary Button, does not work,
+    //                            //Tried adding IconElement into icon Props, can't see it.
+    //                            //Tried adding Icon Element into the div next to Primary button and could see it.
+    //                            //
+    //                                 <div id={ buttonID } title={ buttonTitle } >
+    //                                   <PrimaryButton style= { buttonStyles } iconProps= { buttonIcon } text={b.label} onClick={this._panelButtonClicked.bind(this)} disabled={b.disabled} checked={b.checked} /></div>:
 
-                                      <div id={ buttonID } title={ buttonTitle } >
-                                        <DefaultButton style= { buttonStyles } iconProps= { buttonIcon } text={b.label} onClick={this._panelButtonClicked.bind(this)} disabled={b.disabled} checked={b.checked} /></div>;
+    //                                   <div id={ buttonID } title={ buttonTitle } >
+    //                                     <DefaultButton style= { buttonStyles } iconProps= { buttonIcon } text={b.label} onClick={this._panelButtonClicked.bind(this)} disabled={b.disabled} checked={b.checked} /></div>;
 
-                                buttons.push( thisButton );
-                            }
+    //                             buttons.push( thisButton );
+    //                         }
     
-                        }); //END buttonRow.map( (b,i) => {
+    //                     }); //END buttonRow.map( (b,i) => {
     
-                        const stackQuickCommands: IStackTokens = { childrenGap: 10 };
-                        rowResult = <Stack horizontal={ true } tokens={stackQuickCommands}>
-                            {buttons}
+    //                     const stackQuickCommands: IStackTokens = { childrenGap: 10 };
+    //                     rowResult = <Stack horizontal={ true } tokens={stackQuickCommands}>
+    //                         {buttons}
 
-                        </Stack>;
+    //                     </Stack>;
     
-                        const styleRows: any = {paddingBottom: 10};
-                        if ( quickCommands.styleRow ) {
-                            try {
-                                Object.keys(quickCommands.styleRow).map( k => {
-                                    styleRows[k] = quickCommands.styleRow[k];
-                                });
-                            } catch (e) {
-                                alert( `quickCommands.styleRow is not valid JSON... please fix: ${quickCommands.styleRow}` );
-                            }
-                        }
-                        allButtonRows.push( <div style={ styleRows }> { rowResult } </div> );
+    //                     const styleRows: any = {paddingBottom: 10};
+    //                     if ( quickCommands.styleRow ) {
+    //                         try {
+    //                             Object.keys(quickCommands.styleRow).map( k => {
+    //                                 styleRows[k] = quickCommands.styleRow[k];
+    //                             });
+    //                         } catch (e) {
+    //                             alert( `quickCommands.styleRow is not valid JSON... please fix: ${quickCommands.styleRow}` );
+    //                         }
+    //                     }
+    //                     allButtonRows.push( <div style={ styleRows }> { rowResult } </div> );
 
     
-                    } //END   if ( buttonRow && buttonRow.length > 0 ) {
+    //                 } //END   if ( buttonRow && buttonRow.length > 0 ) {
     
-                }); //END  quickCommands.buttons.map( (buttonRow, r) => {
-                allButtonRows.push( <Icon iconName= { 'Emoji2' } style={ { fontSize: '24px', } }/> );
+    //             }); //END  quickCommands.buttons.map( (buttonRow, r) => {
+    //             allButtonRows.push( <Icon iconName= { 'Emoji2' } style={ { fontSize: '24px', } }/> );
                 
-            } //END   if ( buildAllButtonsTest === true ) {
+    //         } //END   if ( buildAllButtonsTest === true ) {
 
 
-        }
+    //     }
 
-        return allButtonRows;
+    //     return allButtonRows;
 
-    }
+    // }
 
     private covertFieldInfoToIViewFields( parentListFieldTitles: [] , fieldsToShow: string[] ) {
 
@@ -370,6 +374,9 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
 
           firstVisible: 0,
           lastVisible: this.props.itemsPerPage - 1,
+
+          commandResult: null,
+          commandError: false,
         };
     }
         
@@ -465,6 +472,8 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
 
             let dialog = !this.state.myDialog.showDialog ? null : buildConfirmDialogBig( this.state.myDialog );
 
+            const { quickCommands,  } = this.props;
+            const { commandError, commandResult } = this.state;
 
             /***
              *    d888888b  .d88b.   d888b   d888b  db      d88888b .d8888. 
@@ -505,7 +514,8 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
                           { itemLink }
                           <div id='20pxSpacer' style={{ height: '20px'}}/>
                           { attachments }
-                          { this.createPanelButtons( this.props.quickCommands, this.state.panelItem, this.props.sourceUserInfo ) }
+                          {/* { createPanelButtonsV1( this.props.quickCommands, this.state.panelItem, this.props.sourceUserInfo, this._panelButtonClicked.bind(this), this.delim ) } */}
+                          { createPanelButtonsV2( this.props.quickCommands, this.state.panelItem, this.props.sourceUserInfo, this._panelButtonClickedv2.bind(this), this.delim ) }
                           {/* { <Icon iconName= { 'Add' } style={ defaultBannerCommandStyles }/> } */}
                         </div>
                       </PivotItem>
@@ -681,6 +691,21 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
                     </span>
                   </div>;
 
+              /**
+               * Brought this in directly from Drilldown web part
+               */
+              let createBanner = quickCommands !== null && quickCommands.successBanner > 0 ? true : false; //CommandItemNotUpdatedMessage
+              const bannerEleClasses = [ 'quickCommandFooterStyles', commandResult ? 'quickCommandHide' : 'quickCommandShow' ];
+
+              if ( commandResult.status === 'Success' ) {  }
+              else if ( commandResult.status === 'Unknown' ) { bannerEleClasses.push( 'quickCommandWarn' ); }
+              else { bannerEleClasses.push( 'quickCommandError' ); }
+
+              let bannerMessageEle = createBanner === false ? null : <div className={ bannerEleClasses.join(' ') }>
+                  { commandResult.statusElement ? commandResult.statusElement : commandResult.errorInfo.returnMess }
+              </div>;
+
+
               /*stylesL.reactListView*/
               return (
                 <div className={ '' } 
@@ -696,6 +721,7 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
                       { fullPanel }
                       { attachPanel }
                       { dialog }
+                      { bannerMessageEle }
                       { listView }
                     </div>
                 </div>
@@ -797,25 +823,45 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
 
  //private async ensureTrackTimeList(myListName: string, myListDesc: string, ProjectOrTime: string): Promise<boolean> {
      
-    private async _panelButtonClicked ( item: any ): Promise<void> {
+    // private async _panelButtonClicked ( item: any ): Promise<void> {
 
-        let e: any = event;
-        let thisID = findParentElementPropLikeThis(e.target, 'id', 'ButtonID', 5, 'begins');
+    //     let e: any = event;
+    //     let thisID = findParentElementPropLikeThis(e.target, 'id', 'ButtonID', 5, 'begins');
 
-        if ( !thisID ) { 
+    //     if ( !thisID ) { 
 
-            alert('Whoops! Can not find ID of _panelButtonClicked!');
-            return null;
+    //         alert('Whoops! Can not find ID of _panelButtonClicked!');
+    //         return null;
 
-        } else {
+    //     } else {
 
-          try {
-            await this.startThisQuickUpdate( thisID );
-          } catch (ev) {
-            console.log('_panelButtonClicked error:', ev );
-          }
+    //       try {
+    //         await this.startThisQuickUpdate( thisID );
+    //       } catch (ev) {
+    //         console.log('_panelButtonClicked error:', ev );
+    //       }
+    //     }
+    // }
+
+
+    private async _panelButtonClickedv2 ( b: any, item: any ): Promise<void> {
+
+      console.log( '_panelButtonClickedv2', item );
+
+      if ( !item ) { 
+
+          alert('Whoops! Can not find ID of _panelButtonClicked!');
+          return null;
+
+      } else {
+
+        try {
+          await this.startThisQuickUpdate2( b, item );
+        } catch (ev) {
+          console.log('_panelButtonClicked error:', ev );
         }
-    }
+      }
+  }
 
     /**
      * Open the dialog
@@ -844,65 +890,114 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
       this.setState({ fontSize: this._ListViewFontSizes[ nextIdx ] });
     }
 
-    private async startThisQuickUpdate ( thisID: string ): Promise<void>{
+    private async startThisQuickUpdate2 ( thisButtonObject: any, item: any ): Promise<void>{
 
-        const buttonID = thisID.split(this.delim);
-        //let buttonID = ['ButtonID', r, i , item.Id].join(this.delim);
-        const buttonRow: any = buttonID[1];
-        const buttonIndex: any = buttonID[2];
-        const itemId = buttonID[3];
-        const thisButtonObject : IQuickButton = this.props.quickCommands.buttons[ buttonRow ][ buttonIndex ];
+      const itemId = item.Id;
 
-        if ( !thisButtonObject ) {
-            alert('_panelButtonClicked - can not find thisButtonObject - ' + thisID );
-        } else {
+      if ( !thisButtonObject ) {
+          alert('_panelButtonClicked - can not find thisButtonObject - ' + itemId );
+      } else {
 
-            if ( thisButtonObject.updateItem ) {
-                let readyToUpdate = true;
-                if ( !this.props.webURL ) { alert('Missing listWebUrl for quickCommands') ; readyToUpdate = false ; }
-                if ( !this.props.listName ) { alert('Missing listName for quickCommands') ; readyToUpdate = false ; }
+          if ( thisButtonObject.updateItem ) {
+              let readyToUpdate = true;
+              if ( !this.props.webURL ) { alert('Missing listWebUrl for quickCommands') ; readyToUpdate = false ; }
+              if ( !this.props.listName ) { alert('Missing listName for quickCommands') ; readyToUpdate = false ; }
 
-                if ( readyToUpdate === true ) {
-                    //Attempt to update item:
-                    if ( thisButtonObject.confirm && thisButtonObject.confirm.length > 0 ) { 
+              if ( readyToUpdate === true ) {
+                  //Attempt to update item:
+                  if ( thisButtonObject.confirm && thisButtonObject.confirm.length > 0 ) { 
 
-                        let myDialog: IMyBigDialogProps = this.createBlankDialog();
-                        myDialog.title = "Are you sure you want to make this update?";
-                        myDialog.dialogMessage = thisButtonObject.confirm + '';
-                        myDialog.confirmButton = thisButtonObject.label + '';
-                        myDialog.showDialog = true;
-                        myDialog.maxWidth = 600;
+                      let myDialog: IMyBigDialogProps = this.createBlankDialog();
+                      myDialog.title = "Are you sure you want to make this update?";
+                      myDialog.dialogMessage = thisButtonObject.confirm + '';
+                      myDialog.confirmButton = thisButtonObject.label + '';
+                      myDialog.showDialog = true;
+                      myDialog.maxWidth = 600;
 
-                        this.setState({
-                            pickedCommand: thisButtonObject,
-                            myDialog: myDialog as IMyBigDialogProps,
-                        });
+                      this.setState({
+                          pickedCommand: thisButtonObject,
+                          myDialog: myDialog as IMyBigDialogProps,
+                      });
 
-                    } else {
-                        await this.completeThisQuickUpdate ( itemId, thisButtonObject );
+                  } else {
+                      await this.completeThisQuickUpdate ( itemId, thisButtonObject );
 
-                    }
+                  }
 
 
 
-                } else {
-                    //Don't update item:
-                }
-            }
+              } else {
+                  //Don't update item:
+              }
+          }
 
-            if ( thisButtonObject.panelMessage ) {
-                this.setState({
-                    panelMessage: thisButtonObject.panelMessage,
-                });
-            }
-        }
+          if ( thisButtonObject.panelMessage ) {
+              this.setState({
+                  panelMessage: thisButtonObject.panelMessage,
+              });
+          }
+      }
 
-    }
+  }
+
+    // private async startThisQuickUpdate ( thisID: string ): Promise<void>{
+
+    //     const buttonID = thisID.split(this.delim);
+    //     //let buttonID = ['ButtonID', r, i , item.Id].join(this.delim);
+    //     const buttonRow: any = buttonID[1];
+    //     const buttonIndex: any = buttonID[2];
+    //     const itemId = buttonID[3];
+    //     const thisButtonObject : IQuickButton = this.props.quickCommands.buttons[ buttonRow ][ buttonIndex ];
+
+    //     if ( !thisButtonObject ) {
+    //         alert('_panelButtonClicked - can not find thisButtonObject - ' + thisID );
+    //     } else {
+
+    //         if ( thisButtonObject.updateItem ) {
+    //             let readyToUpdate = true;
+    //             if ( !this.props.webURL ) { alert('Missing listWebUrl for quickCommands') ; readyToUpdate = false ; }
+    //             if ( !this.props.listName ) { alert('Missing listName for quickCommands') ; readyToUpdate = false ; }
+
+    //             if ( readyToUpdate === true ) {
+    //                 //Attempt to update item:
+    //                 if ( thisButtonObject.confirm && thisButtonObject.confirm.length > 0 ) { 
+
+    //                     let myDialog: IMyBigDialogProps = this.createBlankDialog();
+    //                     myDialog.title = "Are you sure you want to make this update?";
+    //                     myDialog.dialogMessage = thisButtonObject.confirm + '';
+    //                     myDialog.confirmButton = thisButtonObject.label + '';
+    //                     myDialog.showDialog = true;
+    //                     myDialog.maxWidth = 600;
+
+    //                     this.setState({
+    //                         pickedCommand: thisButtonObject,
+    //                         myDialog: myDialog as IMyBigDialogProps,
+    //                     });
+
+    //                 } else {
+    //                     await this.completeThisQuickUpdate ( itemId, thisButtonObject );
+
+    //                 }
+
+
+
+    //             } else {
+    //                 //Don't update item:
+    //             }
+    //         }
+
+    //         if ( thisButtonObject.panelMessage ) {
+    //             this.setState({
+    //                 panelMessage: thisButtonObject.panelMessage,
+    //             });
+    //         }
+    //     }
+
+    // }
 
     private async completeThisQuickUpdate( itemId: string, thisButtonObject : IQuickButton ): Promise<void> {
 
-        let result = await updateReactListItem( this.props.webURL, this.props.listName, parseInt(itemId), thisButtonObject, this.props.sourceUserInfo, this.state.panelItem );
-
+        const result = await updateReactListItem( this.props.webURL, this.props.listName, parseInt(itemId), thisButtonObject, this.props.sourceUserInfo, this.state.panelItem );
         //If success (result is error message and null by default )
         // if ( result === null && this.props.quickCommands.onUpdateReload === true ) {
 
@@ -915,18 +1010,34 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
             let updates = Object.keys(thisButtonObject.updateItem).map( k => {
                 return k;
             });
-            let bannerMessage: any = <div style={{ marginTop: '5px'}}> { [
-                <h3 key="h3Finished" style={{paddingTop: '10px'}}>Finished updating item [ {itemId} ]</h3>,
-                <div  key="bannerIncluding">Including: { updates.join(', ')} </div>
+
+            result.statusElement = <div style={{ marginTop: '5px'}}> { [
+              <h3 key="h3Finished" style={{paddingTop: '10px'}}>Finished updating item [ {itemId} ]</h3>,
+              <div  key="bannerIncluding">Including: { updates.join(', ')} </div>
             ] }</div>;
 
-            this.props.quickCommands.refreshCallback( bannerMessage, false );
+            this.props.quickCommands.refreshCallback( result, false );
 
         //https://github.com/mikezimm/fps-library-v2/issues/20
         } else if ( result ) {
-            this.props.quickCommands.refreshCallback( result.errorInfo.friendly, true );
-            
+            this.props.quickCommands.refreshCallback( result, true );
         }
+
+        this.setState({ 
+          commandResult: result,
+          commandError: result.status === 'Success' ? false : true,
+        });
+
+        if ( result ) {
+          const delay = result.status !== 'Success' ? 10000 : this.props.quickCommands.successBanner;
+          setTimeout(() => {
+            this.setState({ 
+              commandResult: null,
+              commandError: false,
+             });
+          } , delay);
+        }
+
     }
     /**
      * Close the dialog